Use a class property to provide a fake for testing
